Crate icu_calendar_data

Data for the icu_calendar crate

This data was generated with CLDR version 46.0.0-BETA2, ICU version icu4x/2024-05-16/75.x, and LSTM segmenter version v0.1.0.

Modules

calendar_chinesecache_v1 🔒

calendar_dangicache_v1 🔒

calendar_islamicobservationalcache_v1 🔒

calendar_islamicummalquracache_v1 🔒

calendar_japanese_v1 🔒

calendar_japanext_v1 🔒

datetime_week_data_v1 🔒

datetime_week_data_v2 🔒

Macros

impl_calendar_chinesecache_v1

Implement DataProvider<ChineseCacheV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_calendar_dangicache_v1

Implement DataProvider<DangiCacheV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_calendar_islamicobservationalcache_v1

Implement DataProvider<IslamicObservationalCacheV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_calendar_islamicummalquracache_v1

Implement DataProvider<IslamicUmmAlQuraCacheV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_calendar_japanese_v1

Implement DataProvider<JapaneseErasV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_calendar_japanext_v1

Implement DataProvider<JapaneseExtendedErasV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_week_data_v1

Implement DataProvider<WeekDataV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impl_datetime_week_data_v2

Implement DataProvider<WeekDataV2Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with icu’s _unstable constructors.

impliterable_calendar_chinesecache_v1

Implement IterableDataProvider<ChineseCacheV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_calendar_dangicache_v1

Implement IterableDataProvider<DangiCacheV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_calendar_islamicobservationalcache_v1

Implement IterableDataProvider<IslamicObservationalCacheV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_calendar_islamicummalquracache_v1

Implement IterableDataProvider<IslamicUmmAlQuraCacheV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_calendar_japanese_v1

Implement IterableDataProvider<JapaneseErasV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_calendar_japanext_v1

Implement IterableDataProvider<JapaneseExtendedErasV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_week_data_v1

Implement IterableDataProvider<WeekDataV1Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

impliterable_datetime_week_data_v2

Implement IterableDataProvider<WeekDataV2Marker> on the given struct using the data hardcoded in this file. This allows the struct to be used with DatagenDriver for this key.

make_provider

Marks a type as a data provider. You can then use macros like impl_core_helloworld_v1 to add implementations.